Split files and start unit tests
