+++ amr-wind/wind_energy/actuator/disk/CMakeLists.txt
@@ -1,4 +1,4 @@
 target_sources(${amr_wind_lib_name} PRIVATE
-
   ActuatorDisk.cpp
+  constant_ct_ops.cpp
   )
