+++ unit_tests/wind_energy/actuator/CMakeLists.txt
@@ -3,6 +3,7 @@ target_sources(${amr_wind_unit_test_exe_name} PRIVATE
   test_actuator_flat_plate.cpp
   test_airfoil.cpp
   test_actuator_free_functions.cpp
+  test_disk_constant_ct.cpp
   )
 
 if (AMR_WIND_ENABLE_OPENFAST)
